+++ bundles/framework/myplaces2/request/EditRequestHandler.js
@@ -26,7 +26,7 @@ Oskari.clazz.define('Oskari.mapframework.bundle.myplaces2.request.EditRequestHan
          *      request to handle
          */
         handleRequest: function (core, request) {
-            var sandbox = core.getSandbox();
+            var sandbox = this.sandbox;
             if (request.getName() === 'MyPlaces.EditPlaceRequest') {
                 this._handleEditPlace(sandbox, request);
             } else if (request.getName() === 'MyPlaces.DeletePlaceRequest') {